Follow similar interface class export pattern from before
