+++ test/unit/firebase-namespace.spec.ts
@@ -49,7 +49,7 @@ import {
   setLogFunction,
 } from '@google-cloud/firestore';
 import { InstanceId } from '../../src/instance-id/instance-id';
-import { InstanceIdImpl } from '../../src/instance-id/instance-id-internal';
+import { InstanceIdService } from '../../src/instance-id/instance-id-internal';
 import { ProjectManagement } from '../../src/project-management/project-management';
 import { SecurityRules } from '../../src/security-rules/security-rules';
 import { RemoteConfig } from '../../src/remote-config/remote-config';
@@ -624,7 +624,7 @@ describe('FirebaseNamespace', () => {
     });
 
     it('should return a reference to InstanceId type', () => {
-      expect(firebaseNamespace.instanceId.InstanceId).to.be.deep.equal(InstanceIdImpl);
+      expect(firebaseNamespace.instanceId.InstanceId).to.be.deep.equal(InstanceIdService);
     });
   });
