+++ test/unit/instance-id/instance-id.spec.ts
@@ -25,8 +25,8 @@ import * as chaiAsPromised from 'chai-as-promised';
 import * as utils from '../utils';
 import * as mocks from '../../resources/mocks';
 
-import { InstanceIdImpl as InstanceId } from '../../../src/instance-id/instance-id-internal';
-import { FirebaseInstanceIdRequestHandler } from '../../../src/instance-id/instance-id-request';
+import { InstanceIdService as InstanceId } from '../../../src/instance-id/instance-id-internal';
+import { FirebaseInstanceIdRequestHandler } from '../../../src/instance-id/instance-id-request-internal';
 import { FirebaseApp } from '../../../src/firebase-app';
 import { FirebaseInstanceIdError, InstanceIdClientErrorCode } from '../../../src/utils/error';
 
@@ -118,7 +118,7 @@ describe('InstanceId', () => {
     it('is read-only', () => {
       expect(() => {
         (iid as any).app = mockApp;
-      }).to.throw('Cannot set property app of #<InstanceIdImpl> which has only a getter');
+      }).to.throw('Cannot set property app of #<InstanceId> which has only a getter');
     });
   });
